feat: 2023_09_08 v.1

+++ front/spring_study/src/api/users.js
@@ -4,10 +4,14 @@ export const registerUser = async (user) => {
     return api.post("/user", user);
 };
 
+export const getUserListAll = async () => {
+    return api.get("/user/list");
+}
+
 export const updateUser = async (user) => {
-    return api.put("/user", user);
+    return api.put(`/users/${user.userId}`, user);
 };
 
 export const deleteUser = async (user) => {
-    return api.delete("/user", user);
+    return api.delete(`/user/${user.userId}`, {params: user});
 };
